requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
asyncpg==0.30.0
imbalanced-learn==0.13.0
joblib==1.4.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
pandas==2.2.3
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sklearn-compat==0.1.3
stringzilla==3.12.3
threadpoolctl==3.6.0
typing_extensions==4.13.0
tzdata==2025.2
